separate api endpoint from  main cauze for after adding another endpoints..
